+++ pyobjc-core/Modules/objc/instance-var.h
@@ -16,7 +16,7 @@ extern PyTypeObject PyObjCInstanceVariable_Type;
 #define PyObjCInstanceVariable_Check(obj)                                                \
     PyObject_TypeCheck((obj), &PyObjCInstanceVariable_Type)
 
-extern PyObject* _Nullable PyObjCInstanceVariable_New(char* name);
+extern PyObject* _Nullable PyObjCInstanceVariable_New(const char* name);
 extern int PyObjCInstanceVariable_SetName(PyObject* self, PyObject* name);
 
 #define PyObjCInstanceVariable_IsOutlet(obj) (((PyObjCInstanceVariable*)(obj))->isOutlet)